mention README.md for hackage's haddocker
I think this is all that is necessary to get hackage to render the `README.md` into html. It works here https://hackage.haskell.org/package/streaming for example (the readme is toward the bottom).  I take it that you noticed the links to the github page are to `github.com/githubuser` ?